+++ tradingstrategy/transport/cache.py
@@ -8,7 +8,7 @@ import platform
 import re
 import json
 from importlib.metadata import version
-from typing import Optional, Callable, Set, Union, Collection, Dict
+from typing import Optional, Callable, Set, Union, Collection, Dict, Literal
 import shutil
 import logging
 
@@ -171,6 +171,7 @@ class CachedHTTPTransport:
         start_time: Optional[datetime.datetime] = None,
         end_time: Optional[datetime.datetime] = None,
         max_bytes: Optional[int] = None,
+        candle_type: Literal["candle", "lending_candle"] = "candle",
     ) -> str:
         """Generate the name of the file for holding cached candle data for ``pair_ids``.
         """
@@ -199,6 +200,9 @@ class CachedHTTPTransport:
 
         end_part = end_time.strftime("%Y-%m-%d_%H-%M-%S") if end_time else "any"
 
+        if candle_type == "lending_candle":
+            return f"lending-candles-{time_bucket.value}-between-{start_part}-and-{end_part}-{md5}.parquet"
+            
         return f"candles-jsonl-{time_bucket.value}-between-{start_part}-and-{end_part}-{md5}.parquet"
 
     def purge_cache(self, filename: Optional[Union[str, pathlib.Path]] = None):
@@ -314,7 +318,7 @@ class CachedHTTPTransport:
         return self.get_cached_item(path)
 
     def fetch_lending_reserves_all_time(self) -> pathlib.Path:
-        fname = "lending-reserves-all.json"
+        fname = "lending-reserves-all.parquet"
         cached = self.get_cached_item(fname)
         if cached:
             return cached
@@ -336,23 +340,20 @@ class CachedHTTPTransport:
         candle_type: str = "variable_borrow_apr",
         start_time: Optional[datetime.datetime] = None,
         end_time: Optional[datetime.datetime] = None,
-        max_bytes: Optional[int] = None,
-        progress_bar_description: Optional[str] = None,
     ) -> pd.DataFrame:
-        """Load particular set of the candles and cache the result.
+        """Load particular set of the lending candles and cache the result.
 
-        If there is no cached result, load using JSONL.
+        For the candles format see :py:mod:`tradingstrategy.lending`.
 
-        More information in :py:mod:`tradingstrategy.transport.jsonl`.
-
-        For the candles format see :py:mod:`tradingstrategy.candle`.
-
-        :param pair_ids:
-            Trading pairs internal ids we query data for.
-            Get internal ids from pair dataset.
+        :param reserve_id:
+            Lending reserve's internal id we query data for.
+            Get internal id from lending reserve universe dataset.
 
         :param time_bucket:
-            Candle time frame
+            Candle time frame.
+
+        :param candle_type:
+            Lending candle type.
 
         :param start_time:
             All candles after this.
@@ -361,28 +362,25 @@ class CachedHTTPTransport:
         :param end_time:
             All candles before this
 
-        :param max_bytes:
-            Limit the streaming response size
-
-        :param progress_bar_description:
-            Display on downlood progress bar
-
         :return:
-            Candles dataframe
+            Lending candles dataframe
         """
         cache_fname = self._generate_cache_name(
-            reserve_id, time_bucket, start_time, end_time, max_bytes
+            reserve_id,
+            time_bucket,
+            start_time,
+            end_time,
+            candle_type="lending_candle",
         )
         cached = self.get_cached_item(cache_fname)
 
         if cached:
             full_fname = self.get_cached_file_path(cache_fname)
-            logger.debug("Using cached JSONL data file %s", full_fname)
+            logger.debug("Using cached data file %s", full_fname)
             return pandas.read_parquet(cached)
         
         api_url = f"{self.endpoint}/lending-reserve/candles"
 
-        # reserve_id to other params
         params = {
             "reserve_id": reserve_id,
             "time_bucket": time_bucket.value,
@@ -395,59 +393,19 @@ class CachedHTTPTransport:
         if end_time:
             params["end"] = end_time.isoformat()
 
-        if max_bytes:
-            params["max_bytes"] = max_bytes
-
         resp = self.requests.get(api_url, params=params, stream=True)
 
-        # # Set progress bar start to the first timestamp
-        # if not progress_bar_start and progress_bar_description:
-        #     progress_bar_start = current_ts
-        #     logger.debug("First candle timestamp at %s", current_ts)
-        #     total = progress_bar_end - progress_bar_start
-        #     assert progress_bar_start <= progress_bar_end, f"Mad progress bar {progress_bar_start} - {progress_bar_end}"
-        #     progress_bar = tqdm(desc=progress_bar_description, total=total)
-
+        # TODO: handle error
         candles = resp.json()[candle_type]
 
-        df = pd.DataFrame(candles)
-        df = df.rename(columns={
-            "ts": "timestamp",
-            "o": "open",
-            "h": "high",
-            "l": "low",
-            "c": "close",
-        })
-        df = df.astype(LendingCandle.DATAFRAME_FIELDS)
-
-        # Convert unix timestamps to Pandas
-        df["timestamp"] = pd.to_datetime(df["timestamp"], unit="s")
-
-        # Assume candles are always indexed by their timestamp
-        df.set_index("timestamp", inplace=True, drop=True)
-
-        # Translate the raw compressed keys to our internal
-        # Pandas keys
-        # for key, value in item.items():
-        #     translated_key = mappings[key]
-        #     if translated_key is None:
-        #         # Deprecated/discarded keys
-        #         continue
-
-        #     candle_data[translated_key].append(value)
-
-        # if idx % refresh_rate == 0:
-        #     if last_ts and progress_bar:
-        #         progress_bar.update(current_ts - last_ts)
-        #         progress_bar.set_postfix({"Currently at": datetime.datetime.utcfromtimestamp(current_ts)})
-        #     last_ts = current_ts
+        df = LendingCandle.convert_web_candles_to_dataframe(candles)
 
         # Update cache
-        # path = self.get_cached_file_path(cache_fname)
-        # df.to_parquet(path)
+        path = self.get_cached_file_path(cache_fname)
+        df.to_parquet(path)
 
-        # size = pathlib.Path(path).stat().st_size
-        # logger.debug(f"Wrote {cache_fname}, disk size is {size:,}b")
+        size = pathlib.Path(path).stat().st_size
+        logger.debug(f"Wrote {cache_fname}, disk size is {size:,}b")
 
         return df